Use 'director' tags as artist display if 'artist' tags not found.
In e6ai.net, there are only Directors and not Artists. In the current version, if users view post details or download pictures, the information will not be displayed.

Since the director label category does not exist on other sites, it can be considered that it has no impact on other sites.